requirements.txt
Django==2.0
django-dirtyfields==1.3.1
django-filter==2.0.0.dev1
django-url-filter==0.3.5
djangorestframework==3.7.2
gunicorn==19.7.1
six==1.10.0
PyJWT==1.6.0
